+++ src/main/java/step2/domain/RaceController.java
@@ -1,4 +1,7 @@
-package step2;
+package step2.domain;
+
+import step2.view.InputView;
+import step2.view.ResultView;
 
 import java.util.List;
