+++ src/main/java/step2/domain/RacingGame.java
@@ -1,7 +1,6 @@
-package step2;
+package step2.domain;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
