+++ src/test/java/step2/RacingGameTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step2.domain.Cars;
+import step2.domain.RacingGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
